+++ web-frontend/modules/database/store/view/gallery.js
@@ -27,11 +27,10 @@ export const mutations = {
 }
 
 export const actions = {
-  async fetchInitial({ dispatch, commit }, { viewId, fields, primary }) {
+  async fetchInitial({ dispatch, commit }, { viewId, fields }) {
     const data = await dispatch('fetchInitialRows', {
       viewId,
       fields,
-      primary,
       initialRowArguments: { includeFieldOptions: true },
     })
     await dispatch('forceUpdateAllFieldOptions', data.field_options)