+++ web-frontend/modules/database/viewTypes.js
@@ -170,7 +170,6 @@ export class ViewType extends Registerable {
     context,
     view,
     fields,
-    primary,
     storePrefix = '',
     includeFieldOptions = false
   ) {}
@@ -228,38 +227,21 @@ export class ViewType extends Registerable {
    * via a real time event by another user. It can be used to check if data in an store
    * needs to be updated.
    */
-  rowCreated(
-    context,
-    tableId,
-    fields,
-    primary,
-    values,
-    metadata,
-    storePrefix
-  ) {}
+  rowCreated(context, tableId, fields, values, metadata, storePrefix) {}
 
   /**
    * Event that is called when a row is updated from an outside source, so for example
    * via a real time event by another user. It can be used to check if data in an store
    * needs to be updated.
    */
-  rowUpdated(
-    context,
-    tableId,
-    fields,
-    primary,
-    row,
-    values,
-    metadata,
-    storePrefix
-  ) {}
+  rowUpdated(context, tableId, fields, row, values, metadata, storePrefix) {}
 
   /**
    * Event that is called when a row is deleted from an outside source, so for example
    * via a real time event by another user. It can be used to check if data in an store
    * needs to be updated.
    */
-  rowDeleted(context, tableId, fields, primary, row, storePrefix) {}
+  rowDeleted(context, tableId, fields, row, storePrefix) {}
 
   /**
    * Event that is called when a piece of metadata is updated for a particular row.
@@ -357,11 +339,10 @@ export class GridViewType extends ViewType {
     return 'database-public-grid-view'
   }
 
-  async fetch({ store }, view, fields, primary, storePrefix = '') {
+  async fetch({ store }, view, fields, storePrefix = '') {
     await store.dispatch(storePrefix + 'view/grid/fetchInitial', {
       gridId: view.id,
       fields,
-      primary,
     })
   }
 
@@ -369,14 +350,12 @@ export class GridViewType extends ViewType {
     { store },
     view,
     fields,
-    primary,
     storePrefix = '',
     includeFieldOptions = false
   ) {
     await store.dispatch(storePrefix + 'view/grid/refresh', {
       view,
       fields,
-      primary,
       includeFieldOptions,
     })
   }
@@ -460,7 +439,6 @@ export class GridViewType extends ViewType {
       storePrefix + 'view/grid/updateSearch',
       {
         fields: rootGetters['field/getAll'],
-        primary: rootGetters['field/getPrimary'],
       },
       {
         root: true,
@@ -482,7 +460,6 @@ export class GridViewType extends ViewType {
     { store },
     tableId,
     fields,
-    primary,
     values,
     metadata,
     storePrefix = ''
@@ -491,14 +468,12 @@ export class GridViewType extends ViewType {
       await store.dispatch(storePrefix + 'view/grid/createdNewRow', {
         view: store.getters['view/getSelected'],
         fields,
-        primary,
         values,
         metadata,
       })
       await store.dispatch(storePrefix + 'view/grid/fetchByScrollTopDelayed', {
         scrollTop: store.getters[storePrefix + 'view/grid/getScrollTop'],
         fields,
-        primary,
       })
       store.dispatch(storePrefix + 'view/grid/fetchAllFieldAggregationData', {
         view: store.getters['view/getSelected'],
@@ -510,7 +485,6 @@ export class GridViewType extends ViewType {
     { store },
     tableId,
     fields,
-    primary,
     row,
     values,
     metadata,
@@ -520,7 +494,6 @@ export class GridViewType extends ViewType {
       await store.dispatch(storePrefix + 'view/grid/updatedExistingRow', {
         view: store.getters['view/getSelected'],
         fields,
-        primary,
         row,
         values,
         metadata,
@@ -528,7 +501,6 @@ export class GridViewType extends ViewType {
       await store.dispatch(storePrefix + 'view/grid/fetchByScrollTopDelayed', {
         scrollTop: store.getters[storePrefix + 'view/grid/getScrollTop'],
         fields,
-        primary,
       })
       store.dispatch(storePrefix + 'view/grid/fetchAllFieldAggregationData', {
         view: store.getters['view/getSelected'],
@@ -536,18 +508,16 @@ export class GridViewType extends ViewType {
     }
   }
 
-  async rowDeleted({ store }, tableId, fields, primary, row, storePrefix = '') {
+  async rowDeleted({ store }, tableId, fields, row, storePrefix = '') {
     if (this.isCurrentView(store, tableId)) {
       await store.dispatch(storePrefix + 'view/grid/deletedExistingRow', {
         view: store.getters['view/getSelected'],
         fields,
-        primary,
         row,
       })
       await store.dispatch(storePrefix + 'view/grid/fetchByScrollTopDelayed', {
         scrollTop: store.getters[storePrefix + 'view/grid/getScrollTop'],
         fields,
-        primary,
       })
       store.dispatch(storePrefix + 'view/grid/fetchAllFieldAggregationData', {
         view: store.getters['view/getSelected'],
@@ -584,11 +554,10 @@ class BaseBufferedRowView extends ViewType {
     return {}
   }
 
-  async fetch({ store }, view, fields, primary, storePrefix = '') {
+  async fetch({ store }, view, fields, storePrefix = '') {
     await store.dispatch(`${storePrefix}view/${this.getType()}/fetchInitial`, {
       viewId: view.id,
       fields,
-      primary,
     })
   }
 
@@ -596,13 +565,11 @@ class BaseBufferedRowView extends ViewType {
     { store },
     view,
     fields,
-    primary,
     storePrefix = '',
     includeFieldOptions = false
   ) {
     await store.dispatch(storePrefix + 'view/' + this.getType() + '/refresh', {
       fields,
-      primary,
       includeFieldOptions,
     })
   }
@@ -679,7 +646,6 @@ class BaseBufferedRowView extends ViewType {
     { store },
     tableId,
     fields,
-    primary,
     values,
     metadata,
     storePrefix = ''
@@ -690,7 +656,6 @@ class BaseBufferedRowView extends ViewType {
         {
           view: store.getters['view/getSelected'],
           fields,
-          primary,
           values,
         }
       )
@@ -701,7 +666,6 @@ class BaseBufferedRowView extends ViewType {
     { store },
     tableId,
     fields,
-    primary,
     row,
     values,
     metadata,
@@ -713,7 +677,6 @@ class BaseBufferedRowView extends ViewType {
         {
           view: store.getters['view/getSelected'],
           fields,
-          primary,
           row,
           values,
         }
@@ -721,14 +684,13 @@ class BaseBufferedRowView extends ViewType {
     }
   }
 
-  async rowDeleted({ store }, tableId, fields, primary, row, storePrefix = '') {
+  async rowDeleted({ store }, tableId, fields, row, storePrefix = '') {
     if (this.isCurrentView(store, tableId)) {
       await store.dispatch(
         storePrefix + 'view/' + this.getType() + '/afterExistingRowDeleted',
         {
           view: store.getters['view/getSelected'],
           fields,
-          primary,
           row,
         }
       )
@@ -797,7 +759,6 @@ export class GalleryViewType extends BaseBufferedRowView {
       storePrefix + 'view/gallery/updateSearch',
       {
         fields: rootGetters['field/getAll'],
-        primary: rootGetters['field/getPrimary'],
       },
       {
         root: true,
@@ -871,7 +832,6 @@ export class FormViewType extends ViewType {
     { store },
     view,
     fields,
-    primary,
     storePrefix = '',
     includeFieldOptions = false
   ) {
@@ -913,7 +873,7 @@ export class FormViewType extends ViewType {
     )
   }
 
-  async fetch({ store }, view, fields, primary, storePrefix = '') {
+  async fetch({ store }, view, fields, storePrefix = '') {
     await store.dispatch(storePrefix + 'view/form/fetchInitial', {
       formId: view.id,
     })